+++ inst/tinytest/test-faviconLink.R
@@ -89,13 +89,13 @@ if (at_home()) {
   # Example of a partial URL that gives a 404. Solved by reading the HTML for the
   # base URL.
   expect_identical_xl(
-    faviconLink("http", "amigo.geneontology.org", "/amigo/term/"),
-    "http://amigo.geneontology.org/static/images/go-logo-favicon.ico"
+    faviconLink("https", "amigo.geneontology.org", "/amigo/term/"),
+    "https://amigo.geneontology.org/static/images/go-logo-favicon.ico"
   )
 
   expect_identical_xl(
-    faviconPlease("http://amigo.geneontology.org/amigo/term/"),
-    "http://amigo.geneontology.org/static/images/go-logo-favicon.ico"
+    faviconPlease("https://amigo.geneontology.org/amigo/term/"),
+    "https://amigo.geneontology.org/static/images/go-logo-favicon.ico"
   )
 
   # Example of a protocol-relative URL